+++ publisher-sdk/src/androidTest/java/com/criteo/publisher/integration/StandaloneFunctionalTest.java
@@ -536,11 +536,6 @@ public class StandaloneFunctionalTest {
       }
 
       @Override
-      public void onAdFailedToDisplay(CriteoErrorCode error) {
-        onFailed();
-      }
-
-      @Override
       public void onAdLeftApplication() {
       }