+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 - Breaking changes
   - Remove `Context` parameter from `CriteoInterstitial` constructor
   - Remove `Parcelable` implementation from `AdSize` class
+  - Move `CriteoInterstitialAdDisplayListener#onAdFailedToDisplay(CriteoErrorCode)` to
+   `CriteoInterstitialAdListener#onAdFailedToReceive(CriteoErrorCode)`. This callback is fired
+    when an error happens after requesting an interstitial ad.
 
 # Version 3.9.0